+++ packages/website/src/pages/editor/@host/@owner/@repository/components/Gitfloat.tsx
@@ -192,7 +192,7 @@ export const Gitfloat = () => {
 			tourStepId: "fork-repository",
 		},
 		hasChanges: {
-			text: "local changes",
+			text: "changes",
 			buttontext: "Push",
 			icon: IconPush,
 			onClick: triggerPushChanges,
@@ -236,7 +236,7 @@ export const Gitfloat = () => {
 			if (userIsCollaborator()) {
 				setIsForking(false)
 			}
-		}),
+		})
 	)
 
 	return (