+++ packages/website/src/services/github/src/router.ts
@@ -52,5 +52,5 @@ router.all(
 			console.error("ERROR in github service: ", error)
 			next(error)
 		}
-	},
+	}
 )